Updated the getBootstraps function to include the matching process
getBootstraps now stores the number of times a specific node from all the bootstrap replicated matched with the comparison tree. In addition, there are now two nonexported functions in TreeFunctions. Namely, the splits_func (used to make the nodes dataframe) and the matching function (used to compare the two nodes dataframes).